avoid duplication of Django version in setup.py & tox.ini
The Django version being specified in setup.py was to allow
an override from tox.ini. This has drawbacks:

* duplication of Django version constraint
* inconvenient to conditionally include depending on the python
  version

The Django version is moved to requirements.txt instead and tox.ini
explicitly overrides with pip install instead of adding a dependency
to the deps configuration variable.

Ideally adding a dependency that does not conflict with the content of
the requirements.txt file would be fine. However, pip does not allow
for this and will throw a "Double requirement" error. An issue exists
to resolve this problem https://github.com/pypa/pip/issues/988